Player/MarioRun: (u8)isForceSlip() narrowing at 3 more callers
getSlopeNormalAccele 99.2->99.9, getSlopeSlideAccele 99.2->99.9,
getChangeAngleSpeed 98.0->98.6. Same lever as getSlideStickMult: target
narrows the BOOL isForceSlip() result with clrlwi. r0,r3,24 before the
test, matching the (u8)isForceSlip() convention already used in
MarioMove.cpp. The cast also collapses the frame inflation that the
wide-int test path caused.

+++ src/Player/MarioRun.cpp
@@ -185,7 +185,7 @@ BOOL TMario::isRunningInWater()
 
 void TMario::getSlopeNormalAccele(f32* accelUp, f32* accelDown)
 {
-	if (isForceSlip()) {
+	if ((u8)isForceSlip()) {
 		*accelUp = mSlipParamsAll.mSlopeAcceleUp.value;
 		*accelDown = mSlipParamsAll.mSlopeAcceleDown.value;
 		return;
@@ -238,7 +238,7 @@ void TMario::getSlopeNormalAccele(f32* accelUp, f32* accelDown)
 
 void TMario::getSlopeSlideAccele(f32* accelUp, f32* accelDown)
 {
-	if (isForceSlip()) {
+	if ((u8)isForceSlip()) {
 		*accelUp = mSlipParamsAll.mSlideAcceleUp.value;
 		*accelDown = mSlipParamsAll.mSlideAcceleDown.value;
 		return;
@@ -295,7 +295,7 @@ f32 TMario::getChangeAngleSpeed()
 {
 	f32 angleSpeed;
 
-	if (isForceSlip()) {
+	if ((u8)isForceSlip()) {
 		angleSpeed = (f32) mSlipParamsAll.mSlideAngleYSp.value;
 	} else {
 		const TBGCheckData* ground = mGroundPlane;
